endeavoru: Remove defines and files that are in tegra3-common
EndeavoruParts from enrc2b adds "double touch to wake" (dt2w) and
"calibration control" features. We can support both in our kernel.

We will also use pollyd and jpolly from enrc2b which is slightly different.
If we encounter any problems with in-call volume we have to add the
endeavoru specific parts to it which should be trivial.

The camerawrapper and liblights were identical.

Change-Id: Ie6c35555b5a629d7fee92438bab69e21ccbfc761