endeavoru: Fix weird green LED blinks
Under some circumstances the green LED blinked whenever the button backlight
turned on or off. I believe this was due to the "new" tegra3-commons
liblights writing to the currents file for whatever reason. Fix that by
removing write permissions to that file.
Also remove button-brightness chmod/chgrp because it's set in tegra3-common.

Change-Id: I79ab51cfe649b2441264a0f8b10a83fa74c869ec